refactor #81: 풀이 종료 메서드 Room 도메인으로 이동
